requirements.txt
# 必需依赖
Pillow>=9.5

# 可选依赖（用于 COCO API）
# pycocotools>=2.0.7

